requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
faiss-cpu
python-dotenv
